+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/domain/importpoc/service/PublishServiceImpl.java
@@ -21,6 +21,7 @@ package org.eclipse.tractusx.traceability.assets.domain.importpoc.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.traceability.assets.application.importpoc.PublishService;
+import org.eclipse.tractusx.traceability.assets.domain.asbuilt.exception.AssetNotFoundException;
 import org.eclipse.tractusx.traceability.assets.domain.asbuilt.repository.AssetAsBuiltRepository;
 import org.eclipse.tractusx.traceability.assets.domain.asplanned.repository.AssetAsPlannedRepository;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.AssetBase;
@@ -37,21 +38,33 @@ public class PublishServiceImpl implements PublishService {
     private final AssetAsPlannedRepository assetAsPlannedRepository;
     private final AssetAsBuiltRepository assetAsBuiltRepository;
 
+    private static void checkImportState(AssetBase assetBase) {
+        if (!ImportState.TRANSIENT.equals(assetBase.getImportState())) {
+            throw new AssetNotFoundException("Asset with ID " + assetBase.getId() + " is not in TRANSIENT state.");
+        }
+    }
+
     @Override
     public void publishAssets(String policyId, List<String> assetIds) {
-        //Update assets with policy id
+
         List<AssetBase> assetAsPlannedList = assetAsPlannedRepository.getAssetsById(assetIds).stream()
-                .filter(assetAsPlanned -> ImportState.TRANSIENT.equals(assetAsPlanned.getImportState()))
-                .peek(assetAsPlanned -> assetAsPlanned.setImportState(ImportState.IN_SYNCHRONIZATION))
-                .peek(assetAsPlanned -> assetAsPlanned.setPolicyId(policyId))
+                .peek(assetAsPlanned -> {
+                    checkImportState(assetAsPlanned);
+                    assetAsPlanned.setImportState(ImportState.IN_SYNCHRONIZATION);
+                    assetAsPlanned.setPolicyId(policyId);
+                })
                 .toList();
         assetAsPlannedRepository.saveAll(assetAsPlannedList);
 
         List<AssetBase> assetAsBuiltList = assetAsBuiltRepository.getAssetsById(assetIds).stream()
-                .filter(assetAsBuilt -> ImportState.TRANSIENT.equals(assetAsBuilt.getImportState()))
-                .peek(assetAsBuilt -> assetAsBuilt.setImportState(ImportState.IN_SYNCHRONIZATION))
-                .peek(assetAsBuilt -> assetAsBuilt.setPolicyId(policyId))
+                .peek(assetAsBuilt -> {
+                    checkImportState(assetAsBuilt);
+                    assetAsBuilt.setImportState(ImportState.IN_SYNCHRONIZATION);
+                    assetAsBuilt.setPolicyId(policyId);
+                })
                 .toList();
         assetAsBuiltRepository.saveAll(assetAsBuiltList);
     }
+
+
 }